Implement plugin details modal redesign
